refactor: centralize cn helper
